smartd: Removed static from variables and functions
Moved initialization of command-line arguments from .h to ParseOpts

Moved initialization of signal handlers to Initialize()

Moved initialization of standby_disable_state to main_worker(). Although it
would be more meaningful to use it in RegisterDevices(), but it might get called
from main_worker not only on first pass, but also on sigHUP. To not modify the
behaviour, I initialize it before it enters the loop.

Moved initialization of is_initialized to main/smartd_main()

Renamed Usage() to smartd_usage() not to conflict with the same function
in smartctl.cpp

Renamed main_worker to smartd_main_worker as there's the same function
in smartctl.cpp